requirements.txt
streamlit>=1.28.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0